requirements.txt
flask==3.0.2
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
pymupdf==1.23.26
python-docx==1.1.0
pandas==2.2.1
python-dotenv==1.0.1
flask-session==0.6.0
flask-wtf==1.2.1
email-validator==2.1.0.post1
gunicorn==21.2.0
pytz==2024.1 